Initial commit
I'm a bit fastidious about the quality of my commits, so in the early
stages of a project I avoid keeping a real git history that could slow
me down. Since progress hasn't been great lately either way, I think
it's a good point to get started.

For the record, I do hold on to the old commits locally, but they are
too messy to be made public without a good reason.